feat: extract site setup helpers
