+++ Codex.md
@@ -4,6 +4,7 @@
 
 ## Summary of Recent Changes
 
+2025-09-07: Split site routes into site-create, pitch-setup, and collective-setup modules – shrink monolith and enable focused ownership – Code-Explorer and Card-Builder may need to update any deep imports relying on old site-routes internals.
 2025-09-11: Externalized default slide metadata to seeds with object storage paths – centralize configuration for branding – Replit styling and Card-Builder exports can swap slide assets per deployment.
 2025-09-10: Added boot tests for memory and Postgres storage using Playwright – verify /api/health and lead submission – ensure server boot and persistence across backends.
 2025-09-09: Documented local development modes for Docker Postgres and in-memory setups – clarify environment variables and seed scripts – developers can choose Postgres containers or memory storage.
