+++ server/site-routes/index.ts
@@ -1,201 +1,19 @@
 import type { Express, Response } from "express";
 import express from "express";
 import path from "path";
-import { readFile } from "fs/promises";
-import { fileURLToPath } from "url";
-import { siteStorage, updateSiteLead } from "./site-storage";
-import { storage } from "./storage";
-import { qrGenerator } from "./qr-generator";
-import { submitToHubSpotForm } from "./hubspot";
-import { createHubSpotService, type ContactData } from "./hubspot-service";
-import { ObjectStorageService, ObjectNotFoundError } from "./objectStorage";
+import { siteStorage, updateSiteLead } from "../site-storage";
+import { storage } from "../storage";
+import { qrGenerator } from "../qr-generator";
+import { submitToHubSpotForm } from "../hubspot";
+import { createHubSpotService, type ContactData } from "../hubspot-service";
+import { ObjectStorageService, ObjectNotFoundError } from "../objectStorage";
 import { insertSiteSchema, insertSiteLeadSchema, insertLegalDisclaimerSchema, insertSiteDisclaimerSchema, insertSiteSlideSchema, insertGlobalSlideSchema, insertSiteAnalyticsSchema } from "@shared/site-schema";
-import { checkSiteAccess, requireAdmin } from "./site-access-control";
-import { isAuthenticated } from "./google-auth";
+import { checkSiteAccess, requireAdmin } from "../site-access-control";
+import { isAuthenticated } from "../google-auth";
 import { z } from "zod";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-// Function to create default slides for a site
-async function createDefaultSlides(siteId: string) {
-  const slidesPath = path.join(__dirname, "seeds", "slides.json");
-  const raw = await readFile(slidesPath, "utf-8");
-  const defaultSlides: Array<{ title: string; imageUrl: string; slideOrder: string; description: string }> = JSON.parse(raw);
-  const [publicPath] = new ObjectStorageService().getPublicObjectSearchPaths();
-
-  for (const slide of defaultSlides) {
-    const imageUrl = path.posix.join(publicPath, slide.imageUrl);
-    await siteStorage.createSiteSlide({
-      siteId: siteId,
-      title: slide.title,
-      imageUrl,
-      slideOrder: slide.slideOrder,
-      isVisible: true,
-      slideType: "image",
-      description: slide.description,
-    });
-  }
-}
-
-// Function to set up default form templates for a pitch site
-async function createDefaultPitchSiteSetup(siteId: string) {
-  try {
-    // Get some of the built-in form templates
-    const formTemplates = await storage.getFormTemplates();
-    const builtInTemplates = formTemplates.filter(template => template.isBuiltIn);
-    
-    // Assign a few key templates to the new pitch site
-    const defaultTemplates = [
-      builtInTemplates.find(t => t.name === 'Learn More'),
-      builtInTemplates.find(t => t.name === 'Contact Sales'),
-      builtInTemplates.find(t => t.name === 'Product Demo')
-    ].filter(Boolean);
-
-    for (let i = 0; i < defaultTemplates.length; i++) {
-      const template = defaultTemplates[i];
-      if (template) {
-        await storage.assignFormToSite({
-          siteId: siteId,
-          formTemplateId: template.id,
-          displayOrder: String(i + 1),
-          cardPosition: i < 2 ? 'main' : 'sidebar',
-          isActive: true,
-          overrideConfig: null
-        });
-      }
-    }
-    
-    console.log(`Created default form assignments for pitch site: ${siteId}`);
-  } catch (error) {
-    console.error('Error setting up default pitch site forms:', error);
-  }
-}
-
-// Function to set up default collective setup (sections and join card)
-async function createDefaultCollectiveSetup(siteId: string) {
-  try {
-    // Get the site to access its name for generating defaults
-    const site = await siteStorage.getSite(siteId);
-    const siteName = site?.name || 'Collective';
-    
-    // Set up default collective settings
-    await siteStorage.updateSite(siteId, {
-      collectiveSettings: JSON.stringify({
-        joinType: 'public', // Default to open join
-        visibility: 'visible', // Default to publicly discoverable
-        maxMembers: null, // No member limit by default
-        autoApprove: false,
-        description: 'Welcome to our collective community',
-        welcomeMessage: 'Thanks for joining our collective! We\'re excited to have you as part of our community.'
-      }) as any,
-      isLaunched: false, // Collective sites start as not launched
-      landingConfig: {
-        heroTitle: `Welcome to ${siteName}`,
-        heroSubtitle: 'Join our community and collaborate together',
-        companyName: siteName,
-        formsTitle: 'Join Our Community',
-        formsDescription: 'Become a member of our collective and help shape our shared goals.',
-        // footerText: `© 2025 ${siteName}. Building community together.` // Not in schema
-      }
-    });
-
-    // Automatically add the Join Card template to the collective homepage
-    const joinCardTemplate = await storage.getJoinCardTemplate();
-    if (joinCardTemplate) {
-      await storage.assignFormToSite({
-        siteId: siteId,
-        formTemplateId: joinCardTemplate.id,
-        displayOrder: '1', // Make it the first card
-        cardPosition: 'main',
-        isActive: true,
-        overrideConfig: {
-          title: `Join ${siteName}`,
-          subtitle: 'Become a member of our community',
-          description: 'Click to join this collective and gain access to member-only content and features.'
-        }
-      });
-      console.log(`Added Join Card template to collective site: ${siteId}`);
-    } else {
-      console.warn('Join Card template not found - collective site created without automatic join card');
-    }
-    
-    console.log(`Created default collective setup for site: ${siteId}`);
-  } catch (error) {
-    console.error('Error setting up default collective setup:', error);
-  }
-}
-
+import { registerSiteCreateRoutes } from "./site-create";
 export function registerSiteRoutes(app: Express, storage?: any) {
-  // Create a new site
-  app.post("/api/sites", async (req, res) => {
-    try {
-      const { presentationMode, ...siteData } = req.body;
-      
-      // Validate that siteType is provided
-      if (!siteData.siteType) {
-        return res.status(400).json({ error: "Site type is required to determine how to construct the site" });
-      }
-      
-      const validatedData = insertSiteSchema.parse(siteData);
-      
-      // Check if siteId already exists
-      const existingSite = await siteStorage.getSite(validatedData.siteId);
-      if (existingSite) {
-        return res.status(400).json({ error: "Site ID already exists" });
-      }
-
-      const site = await siteStorage.createSite(validatedData);
-      
-      // Create content based on site type and presentation mode
-      const mode = presentationMode || 'default';
-      
-      if (validatedData.siteType === 'pitch-site') {
-        // For generic pitch sites, set up form library integration
-        if (mode === 'coming-soon') {
-          // Create a coming soon landing page setup
-          console.log(`Pitch site created with coming-soon mode - landing page ready for customization`);
-        } else if (mode === 'configure-now') {
-          await createDefaultPitchSiteSetup(site.siteId);
-          console.log(`Pitch site created with configure-now mode - redirecting to admin panel`);
-        }
-      } else if (validatedData.siteType === 'collective') {
-        // For collective sites, set up sections and member management (no slides)
-        if (mode === 'coming-soon') {
-          // Create a coming soon landing page setup for collective
-          console.log(`Collective site created with coming-soon mode - landing page ready for customization`);
-        } else if (mode === 'configure-now') {
-          await createDefaultCollectiveSetup(site.siteId);
-          console.log(`Collective site created with configure-now mode - redirecting to admin panel`);
-        }
-      } else {
-        // For mining syndicate sites, use the traditional slide approach
-        if (mode === 'default') {
-          // Create full presentation with all default slides
-          await createDefaultSlides(site.siteId);
-        } else if (mode === 'load-later' || mode === 'load-immediately') {
-          // Don't create any content slides - only the final action cards will appear via global slides
-          console.log(`Site created with ${mode} mode - no content slides added`);
-        }
-      }
-      
-      // Generate QR code for the site
-      const siteUrl = `${req.protocol}://${req.get('host')}/site/${site.siteId}`;
-      const qrCodeUrl = await qrGenerator.generateQRCode(siteUrl, site.siteId);
-      
-      // Update site with QR code URL
-      const updatedSite = await siteStorage.updateSite(site.siteId, {
-        qrCodeUrl: qrCodeUrl
-      });
-
-      res.json(updatedSite);
-    } catch (error) {
-      console.error("Error creating site:", error);
-      if (error instanceof z.ZodError) {
-        return res.status(400).json({ error: "Validation error", details: error.errors });
-      }
-      res.status(500).json({ error: "Failed to create site" });
-    }
-  });
+  registerSiteCreateRoutes(app);
 
   // Get site by siteId (no access control for basic site info - admin panel handles its own access control)
   app.get("/api/sites/:siteId", async (req, res) => {
